smart-customer-portal/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Faker>=28.0.0
scikit-learn>=1.5.0
joblib>=1.4.0
openai>=1.0.0
